Fix renamed file breaking ci build
